refactored tests to allow for more database clients
